Update README.md
The instructions were updated to discuss how to create an orphan bench in a repository. This is needed because trying to create a normal feature branch and then pulling gh-pages from the master will result in a fatal error: cannot merge unrelated histories. This is resolved by creating an orphan branch.